Documentation for PlutoReport.jl

Installation

Right now, the package is not published to General repository. Hence, to use this package, you must use the following code in the top cell.

begin
    import Pkg
    Pkg.add(PackageSpec(url="https://github.com/DhruvaSambrani/PlutoReport.jl"))
    using PlutoReport
end

!!! warning Note This will disable Pluto's inbuild Pkg Manager.

Example

You can find an example notebook here with the associated bibtex file here. Place both in the same folder and run the notebook via Pluto.

API Documentation

PlutoReport.PresentationControls — Type

PresentationControls

alias for NamedTuple{(:presentation_mode, :max_time, :clock),Tuple{Bool,Int64,Int64}}, which is used in presentation_ui

See also presentation_ui, presentation_controls

PlutoReport.presentation — Method

presentation(pvals::PresentationControls) :: HypertextLiteral.Result

Returns htl string which turns UI into presentation mode. Must be the last statement in a cell.

See also presentation_ui, presentation_controls, PresentationControls

PlutoReport.presentation_controls — Method

presentation_controls(aside::Bool = true) :: PlutoUI.CombineNotebook.CombinedBonds

Returns a PlutoUI widget which returns a PresentationControls, and acts as a control panel for presentation_ui. Must be bound to a variable using PlutoUI.bind.

Arguments

aside::Bool: Whether control panel is floating

See also presentation_ui, PresentationControls

PlutoReport.presentation_ui — Method

presentation_ui(pvals::PresentationControls) :: HypertextLiteral.Result

Returns htl string with presentation and timeleft Must be the last statement in a cell.

See also presentation_controls, PresentationControls

PlutoReport.timeleft — Method

timeleft(pvals::PresentationControls) :: HypertextLiteral.Result

Returns string which displays the time left in talk. Must be the last statement in a cell.

See also presentation_ui, presentation_controls, PresentationControls

PlutoReport.References — Method

References() :: HypertextLiteral.Result

Returns a PlutoUI widget which updates a Bond with the list of all citations (with repetition) in the report. Use with display_bibliography, with this bound to a variable using @bind

See also display_bibliography, @cite_str

PlutoReport.display_bibentry — Method

display_bibentry(entry::Bibliography.BibInternal.Entry)::HypertextLiteral.Result

Returns a display for a bibentry. Must be the last statement in a cell.

See also display_bibliography

PlutoReport.display_bibliography — Method

display_bibliography(bibtexpath::String, citations) ::HypertextLiteral.Result

Returns all cited references from a reference BibTex file at bibtexpath and citations. Must be the last statement in a cell.

See also @cite_str, References

PlutoReport.@cite_str — Macro

@cite_str -> HypertextLiteral.Result

Create an inline citation from string

Example

cite"test1"

This creates a citation for bibtex entry with id as test1

See also display_bibliography, References

PlutoReport.Title — Method

Title(title::String, subtitle::String, author::String, affiliation::String) :: HypertextLiteral.Result

Makes the title for the report. Must be the last statement in a cell.

See also make_title

PlutoReport.make_title — Function

make_title is an alias for Title